selection and freecam

+++ res/line_shader.vert
@@ -1,6 +1,6 @@
 #version 430 core
 
-layout(location = 0) in vec3 in_position;
+layout(location = 0) in ivec3 in_position;
 
 uniform mat4 u_view_proj;
 uniform mat4 u_model;
